requirements.txt
asgiref==3.5.0
Django==3.2.9
django-ckeditor==6.3.0
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-taggit==3.0.0
phonenumberslite==8.12.48
Pillow==9.0.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
